Extract `on_initialize` tests
